+++ backend/src/routes/stockListRoutes.ts
@@ -1,7 +1,6 @@
 import express from 'express';
 
-import {createStockList, deleteStockList, getStockLists, updateStockEntry, updateStockList, getStockListsWithData} from '../controllers/stockListController';
-
+import {createStockList, deleteStockList, getStockLists, getStockListStats, getStockListsWithData, updateStockEntry, updateStockList} from '../controllers/stockListController';
 import {authHandler} from '../middleware/authHandler';
 
 export const stockListRouter = express.Router();
@@ -17,5 +16,8 @@ stockListRouter.get('/:id', authHandler, getStockLists);
 
 stockListRouter.put('/:id', authHandler, updateStockList);
 
-/* If provide symbol in body then deletes an entry. Otherwise deletes entire list.*/
+/* If provide symbol in body then deletes an entry. Otherwise deletes entire
+ * list.*/
 stockListRouter.delete('/:id', authHandler, deleteStockList);
+
+stockListRouter.get('/stats/:id', authHandler, getStockListStats);
