+++ backend/src/services/portfolioService.ts
@@ -6,9 +6,9 @@ import jwt from 'jsonwebtoken';
 
 import db from '../db/connectDb';
 import {ResponseType} from '../models/response';
+import {getPeriod} from '../utils/getPeriod';
 
 import {StockListService} from './stockListService';
-import { getPeriod } from '../utils/getPeriod';
 
 const stockListService = new StockListService();
 
@@ -91,8 +91,9 @@ export class PortfolioService {
     }
   }
 
-  async transferFunds(user_id: number, id_1: number, id_2: number, amount: number):
-  Promise<ResponseType> {
+  async transferFunds(
+      user_id: number, id_1: number, id_2: number,
+      amount: number): Promise<ResponseType> {
     try {
       if (!user_id || !id_1 || !id_2 || !amount) {
         return {error: {status: 400, message: 'Missing/Invalid parameters.'}};
@@ -106,35 +107,36 @@ export class PortfolioService {
 
       // Check balance
       const res = await db.query(
-        `SELECT cash_account FROM Portfolio WHERE port_id = $1 AND user_id = $2`,
-        [id_1, user_id]
-      );
-  
+          `SELECT cash_account FROM Portfolio WHERE port_id = $1 AND user_id = $2`,
+          [id_1, user_id]);
+
       if (res.rows.length === 0) {
         throw new Error(`Sender portfolio ${id_1} does not exist.`);
       }
-  
-      const senderBalance = parseFloat(res.rows[0].cash_account.replace(/[^0-9.-]+/g,""));
+
+      const senderBalance =
+          parseFloat(res.rows[0].cash_account.replace(/[^0-9.-]+/g, ''));
       if (senderBalance < amount) {
         throw new Error('Insufficient funds.');
       }
-  
+
       // Subtract from sender
       await db.query(
-        `UPDATE Portfolio SET cash_account = cash_account - $1::MONEY WHERE port_id = $2 AND user_id = $3`,
-        [amount, id_1, user_id]
-      );
-  
+          `UPDATE Portfolio SET cash_account = cash_account - $1::MONEY WHERE port_id = $2 AND user_id = $3`,
+          [amount, id_1, user_id]);
+
       // Add to receiver
       await db.query(
-        `UPDATE Portfolio SET cash_account = cash_account + $1::MONEY WHERE port_id = $2 AND user_id = $3`,
-        [amount, id_2, user_id]
-      );
-  
+          `UPDATE Portfolio SET cash_account = cash_account + $1::MONEY WHERE port_id = $2 AND user_id = $3`,
+          [amount, id_2, user_id]);
+
       await db.query('COMMIT');
 
       return {
-        data: {message: 'Update successs!', result: {message: "Transaction success"}}
+        data: {
+          message: 'Update successs!',
+          result: {message: 'Transaction success'}
+        }
       }
     } catch (error: any) {
       return {
@@ -397,130 +399,26 @@ export class PortfolioService {
   }
 
   async getPortfolioStats(user_id: number, port_id: number, period: string):
-    Promise<ResponseType> {
-  try {
-    if (!user_id || !port_id) {
-      return {error: {status: 400, message: 'Missing parameters.'}};
+      Promise<ResponseType> {
+    try {
+      if (!user_id || !port_id) {
+        return {error: {status: 400, message: 'Missing parameters.'}};
+      }
+      const result = await db.query(
+          `SELECT sl_id FROM Portfolio WHERE user_id = $1 AND port_id = $2`,
+          [user_id, port_id]);
+
+      if (result.rowCount == 0) {
+        return {error: {status: 404, message: 'no portfolio found for user'}};
+      }
+
+      const sl_id = result.rows[0].sl_id;
+      return await stockListService.getStockListStats(user_id, sl_id, period);
+
+    } catch (error: any) {
+      return {
+        error: {status: 500, message: error.message || 'internal server error'}
+      };
     }
-
-    let interval: string = getPeriod(period);
-
-    // coeff of variance and beta
-    const result = await db.query(
-      `
-      WITH stocks_in_list AS (
-        SELECT symbol FROM StockOwned so 
-        JOIN StockList sl ON so.sl_id = sl.sl_id
-        JOIN Portfolio p ON sl.sl_id = p.sl_id
-        WHERE p.port_id = $1 AND p.user_id = $2
-      ),
-      latest_date AS (
-        SELECT MAX(timestamp) AS max_date FROM HistoricalStockPerformance
-      ),
-      daily_returns AS (
-        SELECT 
-          symbol,
-          timestamp,
-          (close - LAG(close) OVER (PARTITION BY symbol ORDER BY timestamp)) / 
-          LAG(close) OVER (PARTITION BY symbol ORDER BY timestamp) AS daily_return
-        FROM HistoricalStockPerformance
-        WHERE symbol IN (SELECT symbol FROM stocks_in_list)
-      ),
-      filtered_returns AS (
-        SELECT dr.*
-        FROM daily_returns dr
-        JOIN latest_date ld ON dr.timestamp >= ld.max_date - $3::INTERVAL
-      ),
-      market_returns AS (
-        SELECT 
-          timestamp,
-          AVG(daily_return) AS market_return
-        FROM filtered_returns
-        GROUP BY timestamp
-      ),
-      joined_returns AS (
-        SELECT 
-          f.symbol,
-          f.timestamp,
-          f.daily_return,
-          m.market_return
-        FROM filtered_returns f
-        JOIN market_returns m ON f.timestamp = m.timestamp
-      ),
-      cv_beta AS (
-        SELECT
-          symbol,
-          STDDEV(daily_return) / NULLIF(AVG(daily_return), 0) AS coefficient_of_variance,
-          COVAR_POP(daily_return, market_return) / VAR_POP(market_return) AS beta
-        FROM joined_returns
-        GROUP BY symbol
-      )
-      SELECT * FROM cv_beta;
-
-      `,
-      [port_id, user_id, interval]
-    )
-
-    // Covariance/correlation matrix
-    // Note: matix is in long-form: pairwise results per row
-    // Note 2: Only joins stock pairs with matching timestamps
-    const result_matrix = await db.query(
-      `
-      WITH stocks_in_list AS (
-        SELECT symbol FROM StockOwned so 
-        JOIN StockList sl ON so.sl_id = sl.sl_id
-        JOIN Portfolio p ON sl.sl_id = p.sl_id
-        WHERE p.port_id = $1 AND p.user_id = $2
-      ),
-      latest_date AS (
-        SELECT MAX(timestamp) AS max_date FROM HistoricalStockPerformance
-      ),
-      daily_returns AS (
-        SELECT 
-          symbol,
-          timestamp,
-          (close - LAG(close) OVER (PARTITION BY symbol ORDER BY timestamp)) / 
-          LAG(close) OVER (PARTITION BY symbol ORDER BY timestamp) AS daily_return
-        FROM HistoricalStockPerformance
-        WHERE symbol IN (SELECT symbol FROM stocks_in_list)
-      ),
-      filtered_returns AS (
-        SELECT dr.*
-        FROM daily_returns dr
-        JOIN latest_date ld ON dr.timestamp >= ld.max_date - $3::INTERVAL
-      ),
-      pairwise_returns AS (
-        SELECT 
-          a.symbol AS stock_a,
-          b.symbol AS stock_b,
-          a.daily_return AS return_a,
-          b.daily_return AS return_b
-        FROM filtered_returns a
-        JOIN filtered_returns b 
-          ON a.timestamp = b.timestamp AND a.symbol < b.symbol
-      )
-      SELECT 
-        stock_a,
-        stock_b,
-        COVAR_POP(return_a, return_b) AS covariance,
-        CORR(return_a, return_b) AS correlation
-      FROM pairwise_returns
-      GROUP BY stock_a, stock_b
-      ORDER BY stock_a, stock_b;
-
-      `,
-      [port_id, user_id, interval]
-    )
-
-    return {data: {coeff_and_beta: result.rows, matrix: result_matrix.rows}};
-
-  } catch (error: any) {
-    return {
-      error: {status: 500, message: error.message || 'internal server error'}
-    };
   }
-  }
-
 }
-
-
